// src/types/dummy-product-types.ts
export type DummyProductType = {
  id?: string;
    code: string;
    sku: string;
    product_name: string;
    category: string;
    material: string;
    unit: string;
    qty: number;
    status?: string;
    created_at?: string;
    updated_at?: string;
    created_by?: string;
    supplier?: string;
    location?: string;
    approved_by?: string;
    status_approval?: string;
    destination_transfer: string;
    notes?: string;
  };
  